+++ packages/styleguide/stories/Atoms/Icons/constants.tsx
@@ -28,6 +28,7 @@ export const VENDOR_ICONS = pick(
     'MastercardIcon',
     'VisaIcon',
     'GithubOutlineIcon',
+    'JiraIcon',
     'LinkedinOutlineIcon',
     'TwitterOutlineIcon',
   ].sort()