+++ packages/gamut-kit/CHANGELOG.md
@@ -3,6 +3,46 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+### [0.6.37](https://github.com/Codecademy/gamut/compare/@codecademy/gamut-kit@0.6.36...@codecademy/gamut-kit@0.6.37) (2021-12-20)
+
+**Note:** Version bump only for package @codecademy/gamut-kit
+
+
+
+
+
+### [0.6.36](https://github.com/Codecademy/gamut/compare/@codecademy/gamut-kit@0.6.35...@codecademy/gamut-kit@0.6.36) (2021-12-17)
+
+**Note:** Version bump only for package @codecademy/gamut-kit
+
+
+
+
+
+### [0.6.35](https://github.com/Codecademy/gamut/compare/@codecademy/gamut-kit@0.6.34...@codecademy/gamut-kit@0.6.35) (2021-12-16)
+
+**Note:** Version bump only for package @codecademy/gamut-kit
+
+
+
+
+
+### [0.6.34](https://github.com/Codecademy/gamut/compare/@codecademy/gamut-kit@0.6.33...@codecademy/gamut-kit@0.6.34) (2021-12-15)
+
+**Note:** Version bump only for package @codecademy/gamut-kit
+
+
+
+
+
+### [0.6.33](https://github.com/Codecademy/gamut/compare/@codecademy/gamut-kit@0.6.32...@codecademy/gamut-kit@0.6.33) (2021-12-14)
+
+**Note:** Version bump only for package @codecademy/gamut-kit
+
+
+
+
+
 ### [0.6.32](https://github.com/Codecademy/gamut/compare/@codecademy/gamut-kit@0.6.31...@codecademy/gamut-kit@0.6.32) (2021-12-14)
 
 **Note:** Version bump only for package @codecademy/gamut-kit