+++ packages/monaco/CHANGELOG.md
@@ -3,6 +3,30 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+### [0.5.27](http://github.com///compare/@codecademy/monaco@0.5.26...@codecademy/monaco@0.5.27) (2021-12-17)
+
+**Note:** Version bump only for package @codecademy/monaco
+
+
+
+
+
+### [0.5.26](http://github.com///compare/@codecademy/monaco@0.5.25...@codecademy/monaco@0.5.26) (2021-12-16)
+
+**Note:** Version bump only for package @codecademy/monaco
+
+
+
+
+
+### [0.5.25](http://github.com///compare/@codecademy/monaco@0.5.24...@codecademy/monaco@0.5.25) (2021-12-15)
+
+**Note:** Version bump only for package @codecademy/monaco
+
+
+
+
+
 ### [0.5.24](http://github.com///compare/@codecademy/monaco@0.5.23...@codecademy/monaco@0.5.24) (2021-12-14)
 
 **Note:** Version bump only for package @codecademy/monaco